requirements/requirements-cpu.txt
torch==1.10.2+cpu
torchvision==0.11.3+cpu
torchaudio==0.10.2+cpu

mmcv-full==1.4.6
mmdet==2.22.0
mmaction2==0.21.0